+++ flink/v1.14/flink/src/main/java/com/netease/arctic/flink/read/source/log/kafka/LogKafkaRecordEmitter.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package com.netease.arctic.flink.read.source.log;
+package com.netease.arctic.flink.read.source.log.kafka;
 
 import org.apache.flink.api.connector.source.SourceOutput;
 import org.apache.flink.connector.kafka.source.reader.KafkaRecordEmitter;
